+++ Housing.hpp
@@ -3,36 +3,36 @@ class Housing
 	private:
 		double longitude; 
 		double latitude; 
-		int housing_median_age; 
-		int total_rooms; 
-		int total_bedrooms; 
-		int population; 
-		int households; 
+		double housing_median_age; 
+		double total_rooms; 
+		double total_bedrooms; 
+		double population; 
+		double households; 
 		double median_income; 
 
 	public:  
 		Housing();
 		Housing(double longitude, double latitude, 
-		int housing_median_age, 
-		int total_rooms, 
-		int total_bedrooms, 
-		int population,
-		int households, 
+		double housing_median_age, 
+		double total_rooms, 
+		double total_bedrooms, 
+		double population,
+		double households, 
 		double median_income); 
 
 		double get_longitude(); 
 
 		double get_latitude(); 
 
-		int get_housing_median_age(); 
+		double get_housing_median_age(); 
 		
-		int get_total_rooms(); 
+		double get_total_rooms(); 
 
-		int get_total_bedrooms();  
+		double get_total_bedrooms();  
 		
-		int get_population(); 
+		double get_population(); 
 		
-		int get_households();  
+		double get_households();  
 		
 		double get_median_income();
